// Hackathon-Frontend/src/services/openapi/models/CheckUpInformationDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CheckUpInformationDto = {
    id: string;
    patientId: string;
    reason: string;
    dateVisited: string;
    followUpDate: string;
    hospitalId: string;
    hospitalName?: string;
};