// Hackathon-Frontend/src/services/openapi/models/DiagnosisInformationDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type DiagnosisInformationDto = {
    id: string;
    patientId: string;
    hospitalId: string;
    date: string;
    diagnosisName: string;
    result: string;
    hospitalName?: string;
};