// Hackathon-Frontend/src/services/openapi/models/MedicationDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type MedicationDto = {
    id: string;
    hospitalId: string;
    patientId: string;
    date: string;
    medicineName: string;
    hospitalName?: string;
    frequency: string;
};