// Hackathon-Frontend/src/services/openapi/models/PatientInformationDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PatientInformationDto = {
    patientId: string;
    firstName: string;
    lastName: string;
    dateOfBirth: string;
    gender: string;
    phoneNumber: string;
    citizenshipNumber: string;
    bloodGroup: string;
};